+++ packages/core/lib/src/internal/core_ops.dart
@@ -24,6 +24,7 @@ part 'ops/priorities.dart';
 part 'ops/style_background.dart';
 part 'ops/style_border.dart';
 part 'ops/style_display.dart';
+part 'ops/style_ellipsis.dart';
 part 'ops/style_margin.dart';
 part 'ops/style_padding.dart';
 part 'ops/style_sizing.dart';
@@ -40,7 +41,6 @@ part 'ops/tag_pre.dart';
 part 'ops/tag_q.dart';
 part 'ops/tag_ruby.dart';
 part 'ops/tag_table.dart';
-part 'ops/text_style.dart';
 
 const kAttributeId = 'id';
 
@@ -70,11 +70,19 @@ const kTagSub = 'sub';
 const kTagSup = 'sup';
 const kTagTt = 'tt';
 
+const kCssDirection = 'direction';
+const kCssDirectionLtr = 'ltr';
+const kCssDirectionRtl = 'rtl';
+const kAttributeDir = 'dir';
+
 const kCssDisplay = 'display';
 const kCssDisplayBlock = 'block';
 const kCssDisplayInline = 'inline';
 const kCssDisplayInlineBlock = 'inline-block';
 const kCssDisplayNone = 'none';
+
+const kCssLineHeight = 'line-height';
+const kCssLineHeightNormal = 'normal';
 
 const kCssWhitespace = 'white-space';
 const kCssWhitespaceNormal = 'normal';
